Extract executale graph and remove tremor script operator
